@@ -16,6 +16,6 @@ RSpec.configure do |config|
   config.disable_monkey_patching!
   config.infer_spec_type_from_file_location!
 
-  config.include FactoryGirl::Syntax::Methods
+  config.include FactoryBot::Syntax::Methods
   config.include Rails.application.routes.url_helpers
 end
