@@ -1,4 +1,4 @@
-//= require require_config.js.erb
+//= require require_config.js
 
 // Components
 require(['jquery', 'componentLoader', 'eventsWithPromises'], function ($, componentLoader, eventsWithPromises) {
